Add an example of a scala class.
This should look fairly familiar to just about any java developer.
public is the default access specifier in scala, so we can leave
that off when defining public fields. Otherwise, we use private
to make fields not public.

The method add() does nto return a value, as indicated by its return
type of Unit.  Also note that parameters passed to scala methods
are always vals, so attempting to reassign their values is a
compile-time error.