Shorten ChecksumAccumulator's methods.
We can leave off the Unit return type and equals sign and give
add() a procedure-like definition instead.

Methods that only calculate a single value don't need a return,
so we can remove that from checksum().

Finally, the bodies of both methods are small enough to fit on
one line each. In fact, for checksum(), which only has a single
expression, we can even leave off the curly braces.