Add the ChecksumAccumulator companion object.
When a singleton object shares the same name with a class,
it's called that class's companion object. You must define
both the class and its companion object in the same source
file. The class is called the companion class of the singleton
object. A class and its companion object can access each other's
private members.