Use the App trait to save some typing.
As of Scala 2.9, the App trait replaces the deprecated Application
trait.

To use the trait, you first write "extends App" after the name of
your singleton object. Then instead of writing a main method, you
place the code you would have put in the main method directly
between the curly braces of the singleton object.

That's it.  Such applications compile and run just like any other.
The way this works is that the trait App declares a main method
of the appropriate signature, which your singleton object inherits,
making it usable as a Scala application. The code between the curly
braces is collected into a primary constructor of the singleton
object, and is executed when the class is initialized.