Add an example of imperative-style if statement.
This code declares a variable, filename, and initializes it to
a default value. It then uses an if expression to check whether any
arguments were supplied to the program. If so, it changes the
variable to hold the value specified in the argument list. If no
arguments were supplied, it leaves the variable set to the default
value.

This code can be written more nicely, because `if` is an expression
in Scala that results in a value.