Rewrite the if expression in a more functional style.
This time, the if has two branches. If args is not empty, the
initial element, args(0), is chosen. Else, the default value is
chosen. The if expression results in the chosen value, and the
filename variable is initialized with that value. The code is
slightly shorter, but the real advantage is that it uses a val
instead of a var.

Using a val is the functional style, and it helps you in much the
same way as a final variable in Java. It tells readers of the code
that the variable will never change, saving them from scanning all
code in the variable's scope to see if it ever changes.