Use a for expression to print files in the current directory.
The "file <- filesHere" syntax is also called a generator, and it
iterates over filesHere, which is of type Array[File]. In each
iteration, a new val named file is initialized with an element
value. The compiler infers the type of file to be File because
filesHere is an Array[File]. For each iteration, the body of the
expression, println(file), will be executed. Because File's
toString method yields the name of the file or directory,
the names of all the files and directories in the current
directory will be printed.