Use for expressions to iterate over Range types.
"1 to 4" generates a range from 1 to 4, inclusive of 4.
"1 until 4" generates a range from 1 to 4, excluding 4.

Iterating through integers like this is common in Scala, but not
nearly as much as in other languages. In other languages,
you might use this facility to iterate through an array,
like this:

// Not common in Scala...
for (i <- 0 to filesHere.length - 1)
  println(filesHere(i))

This for expression introduces a variable i, sets it in turn to
each integer between 0 and filesHere.length - 1, and executes
the body of the for expression for each setting of i. For each
setting of i, the i’th element of filesHere is extracted
and processed.  The reason this kind of iteration is less common
in Scala is that you can just as well iterate over the collection
directly. If you do, your code becomes shorter and you sidestep
many of the off-by-one errors that can arise when iterating
through arrays. Should you start at 0 or 1? Should you add -1, +1,
or nothing to the final index? Such questions are easily answered,
but easily answered wrongly.

It is safer to avoid such questions entirely.