Catching exceptions in Scala has a slightly different syntax.
The syntax for catch clauses was chosen for its consistency with
an important part of Scala - pattern matching.

The behavior of this try-catch expression is the same as in other
languages with exceptions. The body is executed, and if it throws
an exception, each catch clause is tried in turn.

In this examle, if the exception is of type FileNotFoundException,
the first clause will excecute. If it is of type IOException, the
second clause will execute. If the exception is of neither type,
the try-catch will terminate and the exception will propogate
further.

Unlike java, Scala doesn't require you to catch checked exceptions,
or declare them in a throws clause. You can declare a throws
clause if you wish with the @throws annotation, but it is not
required.