Add an example of the match expression.
Here the match expression looks and feels like the switch
statement from other languages. The default case is matched by an
underscore, a wildcard symbol frequently used in Scala as a
placeholder for a completely unknown value.

Any kind of constant, as well as other things, can be used in cases
in Scala, not just integer types and enum constants as in Java.
Also, there are no breaks, as the breaks are implicit; fallthrough
from one alternative to the next is not possible. The common case -
not falling through - becomes shorter, and a source of errors is
avoided because programmers can no longer fall through by accident.

Finally, match expressions result in a value.

This particular match expression has side effects.