Replace a while loop with recursion.
Every var should be considered a candidate for conversion to a val.
Every iterative loop is suspect, and vars and loops can be replaced
with vals and recursion.

This version of the code gives an intent-revealing name to what
the function does, and it uses recursion to substitute for looping.
Each skipped iteration is replaced with a recursive call with
i + 1 as the argument, effectively skipping to the next integer.

Once people can get their minds wrapped around recursion, they'll
find this style of coding easier to understand.