Add an example of breaks in Scala if you really need them.
You can exit to an enclosing block that's marked with breakable
with the break method offered by scala.control.util.Breaks.

The Breaks class implements break by throwing an exception that is
caught by an enclosing application of the breakable method.
Therefore, the call to break does not need to be in the same
method as the call to breakable.