Function literals can be shortened when applied to a collection.
Obviously, the parameters must match the type of the collection's
elements - Scala lets you leave out such redundant information and
write function literals more briefly.

This is called target typing, because the targeted usage of an
expression - in this case an argument to numbers.filter() - is
allowed to influence the typing of that expression - in this case
to determine the type of the x parameter.

You can simply start by writing a function literal without the
argument type, and, if the compiler gets confused,
add in the type. With time you'll develop a sense for which
situations the compiler can or can't puzzle out.