Add an example of repeated function parameters.
Scala allows you to indicate that the last parameters to a function
may be repeated. This allows clients to pass variable length argument
lists to the function. To denote a repeated parameter, place an
asterisk after the type of the parameter.

The echo() function can be called with zero to many String arguments.

Inside the function, the type of the repeated parameter is an Array of
the declared type of the parameter. Thus, the type of args inside the
echo function, which is declared as type "String*"
is actually Array[String].