Reduce redundant parameter passing.
The filesMatching method was doing nothing with the query parameter,
except to pass it back to matcher, which came from a context which
was already known to the caller!

We can then reduce filesMatching to just take a matcher that just
takes a file name and returns a boolean indicating the match.
Each matcher passed already knows its own query, and filesMatching
need not be concerned with it.