Reduce Element.beside to a single expression.
The zip function yields an array of Tuple2's of the elements
of its operands.  Any elements in the first array that don't
have a match on the second array (suppose the first has 3
elements but the second only 2) are dropped.

The for expression uses pattern matching to destructure the
resulting array of pairs into the values line1 for the first
element and line2 for the second element of each pair in the array.