Add an example of how private access in Scala differs from Java.
In Java classes have access to private members of their inner classes,
whereas in Scala they do not.

However, it is alright for InnerMost to access Inner's f(), because
it's happening from within Inner.

So, code can "reach up" to access private members (as long as it's from
within the class declaring them) but cannot "reach down" to access them,
because the access would then be from outside of the class declaring the
private member, regardless of whether or not the attempt was made from a
containing outer class.