Add an example of how protected access differs in Scala from Java.
Access to protected members is also a bit more restrictive than in Java.
In Scala protected members are only accessible from subclasses of the
class in which the member is defined. In Java such accesses are also
possible from other classes in the same package.

Here, the access to f in class Sub is ok because f is declared protected
in Super and Sub is a subclass of Super. By contrast, the access to f in
Other is not permitted, because Other does not inherit from Super. In
Java, the latter access would still be permitted because Other is in the
same package as Super.