Add an example of augmenting access scope with qualifiers.
Access modifiers in Scala can be augmented with qualifiers. A modifier
of the form private[X] or protected[X] means that access is privatge or
protected "up to" X, where X designates some enclosing package, class or
singleton object.

Qualified access modifiers give you very fine-grained control over
visibility. In particular they enable you to express Java's
accessibility notions such as package private, package protected, or
private up to outermost class, which are not directly expressible with
simple modifiers in Scala.

But they also let you express accessibility rules that cannot be
expressed in Java. In this example, class Navigator is labeled
private[bobsrockets]. This means that this class is visible in all
classes and objects that are contained in package bobsrockets. In
particular, the access to Navigator in object Vehicle is permitted,
because Vehicle is contained in package launch, which is contained in
bobsrockets. On the other hand, all code outside the package bobsrockets
cannot access class Navigator.

This technique is quite useful in large projects that span several
packages. It allows you to define things that are visible in several
sub-packages of your project but that remain hidden from clients
external to your project. The same technique is not possible in Java.
There, once a definition escapes its immediate package boundary, it is
visible to the world at large.