Move the showFruit function from the Fruit object to the bobsdelights package.
This is an example of package objects, which are available on a
one-per-package basis to every package in Scala. Definitions in a
package object are available to all other members of that package.

Any definitions placed in a package object are considered members of the
package itself. Any other code in any package can import the showFruit
utility method just like it would import a class.