Add an example of using the assetion method, ensuring().
This method is another excellent way to perform assertions. Typically,
you'd store the result in a val, performing an assertion on the result,
then mentioning the val last so the result is returned if the assertion
succeeds.

However, the ensuring method does this much more concisely. It appears
to get invoked on the returned result, but actually it takes a function
which accepts Any and returns a boolean.  The anonymous function is what
accepts the Element, and it returns whether the Element's width is less
than or equal to w.