Add an example of a test extending FunSuite.
The "Fun" in FunSuite stands for function. "test" is a method
defined in FunSuite, which will be invoked by the primary
constructor of ElementSuite.

You specify the name of the test as a string between the
parentheses, and the test code itself between curly braces. The test
code is a function passed as a by-name parameter to test, which
registers it for later execution.

One benefit of FunSuite is you need not name all your tests starting
with "test". In addition, you can more easily give long names to
your tests, because you need not encode them in camel case, as you
must do with test methods.