Add another example of wildcard patterns.
Wildcards can also be used to ignore parts of an object that you do not
care about. For example, here we don't care what the elements of the
binary operation are. It just checks whether it is a binary operation at
all.

Thus the code can just as well use the wildcard pattern for the elements
of the BinaryOperator, as shown here.