Add an example of a match on constant patterns.
A constant pattern matches only itself. Any literal may be used as a
constant. For example, 5, true, and "hello" are all constant patterns.
Also, any val or singleton object can be used as a constant. For
example, Nil, a singleton object, is a pattern that matches only the
empty list.