Add an example of a match on variable patterns.
A variable pattern matches any object, just like a wildcard. Unlike a
wildcard, Scala binds the variable to whatever the object is. You can
then use this variable to act on the object further. For example, this
pattern match has a special case for zero, and a default case for all
other values. The default case uses a variable pattern so that it has
a name for the value, no matter what it is.