Add an example of constant pattern matches that can appear to be variable.
Constant patterns can have symbolic names, just like Nil can be used
as a pattern. Here is a related example, where a pattern match
involves the constants E (2.71828. . . ) and PI (3.14159. . . ).

As expected, E does not match PI, so the "strange math" case is not
used. How does the Scala compiler know that PI is a constant imported
from scala.math, and not a variable that stands for the selector value
itself?

Scala uses a simple lexical rule for disambiguation: a simple name
starting with a lowercase letter is taken to be a pattern variable; all
other references are taken to be constants.