Add an example that demonstrates the power of variable pattern matching.
Sometimes this power can get newer users into trouble.  At first glance,
this code appears to desire a pattern match on the constant object, (at
least for the purposes of the match expression; it won't change it) pi.

In fact, actually, pi inside the match is treated as a variable pattern
that catches all values; the compiler graciously allows you to add an
additional case afterward, but warns you of unreachable code.