Add an example showing how backticks can override a variable pattern.
... to behave like a constant pattern.

As you can see, the back-tick syntax for identifiers is used for two
different purposes in Scala to help you code your way out of unusual
circumstances. Here you see that it can be used to treat a lowercase
identifier as a constant in a pattern match.

It can also be used to treat a keyword as an ordinary identifier,
e.g., writing Thread.`yield`() treats yield as an identifier rather
than a keyword.