Add an example of a sequence pattern with an unspecified number of pattern elements.
If you want to match against a sequence without specifying how long
it can be, you can specify _* as the last element of the pattern.

This funny-looking pattern matches any number of elements within a
sequence, including zero elements. Here is an example that matches
any list that starts with zero, regardless of how long the list is.