Add an example of typed pattern matches.
These typed patterns are a convenient replacement for type tests and
type casts.

Here, the generalSize method returns the size or length of objects of
various types. Its argument is of type Any, so it could be any value. If
the argument is a String, the method returns the string's length. The
pattern "s: String" is a typed pattern; it matches every (non-null)
instance of String. The pattern variable s then refers to that string.
The same goes for "m: Map[_, _]". Any non-null Map instance will be
matched by m.