Add an example of what the typed pattern match code would look like without pattern matching.
It's more code, more repetition, less elegant.  Writing type tests
and casts is rather verbose in Scala. That’s intentional, because
it is not encouraged practice. You are usually better off using a
pattern match with a typed pattern. That’s particularly true if
you need to do both a type test and a type cast, because both
operations are then rolled into a single pattern match.