Add an example of variable binding pattern matches.
In addition to the standalone variable patterns, you can also
add a variable to any other pattern. You simply write the variable
name, an at sign (@), and then the pattern. This gives you a
variable-binding pattern. The meaning of such a pattern is to
perform the pattern match as normal, and if the pattern succeeds,
set the variable to the matched object just as with a simple
variable pattern.

As an example, this code shows a pattern match that looks for
the absolute value operation being applied twice in a row.
Such an expression can be simplified to only take the absolute
value one time.