Add a pattern match example where case order matters.
In this example, it is important that the catch-all cases come
after the more specific simplification rules. If you wrote them
in the other order, then the catch-all case would be run in favor
of the more specific rules. In many cases, the compiler will even
complain if you try.