Make Expression a sealed class.
Whenever you write a pattern match, you need to make sure you have
covered all of the possible cases. Sometimes you can do this by adding
a default case at the end of the match, but that only applies if there
is a sensible default behavior. What do you do if there is no default?

How can you ever feel safe that you covered all the cases?  In fact,
you can enlist the help of the Scala compiler in detecting missing
combinations of patterns in a match expression. To be able to do this,
the compiler needs to be able to tell which are the possible cases.

In general, this is impossible in Scala, because new case classes
can be defined at any time and in arbitrary compilation units. For
instance, nothing would prevent you from adding a fifth case class to
the Expression class hierarchy in a different compilation unit from
the one where the other four cases are defined.

The alternative is to make the superclass of your case classes sealed.
A sealed class cannot have any new subclasses added except the ones in
the same file. This is very useful for pattern matching, because it
means you only need to worry about the subclasses you already know about.
What’s more, you get better compiler support as well. If you match
against case classes that inherit from a sealed class, the compiler will
flag missing combinations of patterns with a warning message.

Therefore, if you write a hierarchy of classes intended to be pattern
matched, you should consider sealing them. Simply put the sealed keyword
in front of the class at the top of the hierarchy. Programmers using
your class hierarchy will then feel confident in pattern matching
against it.

The sealed keyword, therefore, is often a license to pattern match.