Add an example of using the @unchecked annotation on pattern match selector expressions.
Use this when you're pattern matching on some types of a sealed class
hierarchy, but not all of them, but you're confident that only the types
you're matching on will ever be passed.

Here, the @unchecked annotation suppresses the exhaustiveness checking
normally done on pattern match expressions. Use it with care.