Add another example of a pattern match in a for expression.
The interesting bit about this example is that even though there
are three elements in the results list, only apple and orange
are printed.

Generated values that do not match the pattern are discarded.
The second element None in the results list does not match the
pattern Some(fruit); therefore it does not show up in the output.