Add SlowHeadQueue, which implements FunctionalQueue in terms of a reversed List.
Now enqueue is constant time, but head and tail are not. They now take
time proportional to the number of elements stored in the queue.

Looking at SlowHeadQueue and SlowAppendQueue, it does not seem easy
to come up with an implementation that’s constant time for all three
operations.

In fact, it looks doubtful that this is even possible!