Hide the entire implementation behind a BasicQueueImplementation private class.
Private constructors and private members are one way to hide the
initial- ization and representation of a class. Another, more radical
way is to hide the class itself and only export a trait that reveals the
public interface of the class.

The code here implements this design. All three methods of the
FunctionalQueue trait are implemented in a subclass
BasicQueueImplementation, which is itself a private inner class of
object BasicQueue. This exposes to clients the same information
as before, but using a different technique.

Instead of hiding individual constructors and methods, this version
hides the whole implementation class.