Add an example of how abstract parameter-less defs can be overridden with vals.
The implementation requirements of defs are less strict than with vals.
Vals need to always return a stable, immutable value. Defs are not
guaranteed to do so.

However, it is possible to override abstract parameter-less defs with
vals, because defs just need to always provide a value, which vals also
do. Defs make no promise of the stability or immutability of their
return values.